+++ src/parser/FormulaParser.ts
@@ -95,6 +95,9 @@ import {
   StringLiteral,
   TimesOp,
 } from './LexerConfig'
+import {RowAddress} from './RowAddress'
+import {ColumnAddress} from './ColumnAddress'
+import {AddressWithSheet} from './Address'
 
 export interface FormulaParserResult {
   ast: Ast,
@@ -145,14 +148,14 @@ export class FormulaParser extends EmbeddedActionsParser {
    * Rule for procedure expressions: SUM(1,A1)
    */
   private procedureExpression: AstRule = this.RULE('procedureExpression', () => {
-    const procedureNameToken = this.CONSUME(ProcedureName) as IExtendedToken
+    const procedureNameToken = this.CONSUME(ProcedureName) as ExtendedToken
     const procedureName = procedureNameToken.image.toUpperCase().slice(0, -1)
     const canonicalProcedureName = this.lexerConfig.functionMapping[procedureName] ?? procedureName
     const args: Ast[] = []
 
     let argument = this.SUBRULE(this.booleanExpressionOrEmpty)
     this.MANY(() => {
-      const separator = this.CONSUME(this.lexerConfig.ArgSeparator) as IExtendedToken
+      const separator = this.CONSUME(this.lexerConfig.ArgSeparator) as ExtendedToken
       if (argument.type === AstNodeType.EMPTY) {
         argument.leadingWhitespace = separator.leadingWhitespace?.image
       }
@@ -166,11 +169,11 @@ export class FormulaParser extends EmbeddedActionsParser {
       args.length = 0
     }
 
-    const rParenToken = this.CONSUME(RParen) as IExtendedToken
+    const rParenToken = this.CONSUME(RParen) as ExtendedToken
     return buildProcedureAst(canonicalProcedureName, args, procedureNameToken.leadingWhitespace, rParenToken.leadingWhitespace)
   })
   private namedExpressionExpression: AstRule = this.RULE('namedExpressionExpression', () => {
-    const name = this.CONSUME(NamedExpression) as IExtendedToken
+    const name = this.CONSUME(NamedExpression) as ExtendedToken
     return buildNamedExpressionAst(name.image, name.leadingWhitespace)
   })
   /**
@@ -189,75 +192,66 @@ export class FormulaParser extends EmbeddedActionsParser {
     this.CONSUME(RParen)
     return this.handleOffsetHeuristic(args)
   })
+
   /**
    * Rule for column range, e.g. A:B, Sheet1!A:B, Sheet1!A:Sheet1!B
    */
   private columnRangeExpression: AstRule = this.RULE('columnRangeExpression', () => {
-    const range = this.CONSUME(ColumnRange) as IExtendedToken
+    const range = this.CONSUME(ColumnRange) as ExtendedToken
     const [startImage, endImage] = range.image.split(':')
+    const firstAddress = this.ACTION(() => columnAddressFromString(this.sheetMapping, startImage, this.formulaAddress))
+    const secondAddress = this.ACTION(() => columnAddressFromString(this.sheetMapping, endImage, this.formulaAddress))
 
-    const start = this.ACTION(() => {
-      return columnAddressFromString(this.sheetMapping, startImage, this.formulaAddress)
-    })
-    let end = this.ACTION(() => {
-      return columnAddressFromString(this.sheetMapping, endImage, this.formulaAddress)
-    })
-
-    if (start === undefined || end === undefined) {
+    if (firstAddress === undefined || secondAddress === undefined) {
       return buildCellErrorAst(new CellError(ErrorType.REF))
     }
-    if (start.exceedsSheetSizeLimits(this.lexerConfig.maxColumns) || end.exceedsSheetSizeLimits(this.lexerConfig.maxColumns)) {
+
+    if (firstAddress.exceedsSheetSizeLimits(this.lexerConfig.maxColumns) || secondAddress.exceedsSheetSizeLimits(this.lexerConfig.maxColumns)) {
       return buildErrorWithRawInputAst(range.image, new CellError(ErrorType.NAME), range.leadingWhitespace)
     }
-    if (start.sheet === undefined && end.sheet !== undefined) {
+
+    if (firstAddress.sheet === undefined && secondAddress.sheet !== undefined) {
       return this.parsingError(ParsingErrorType.ParserError, 'Malformed range expression')
     }
 
-    const sheetReferenceType = this.rangeSheetReferenceType(start.sheet, end.sheet)
+    const { firstEnd, secondEnd, sheetRefType } = FormulaParser.fixSheetIdsForRangeEnds(firstAddress, secondAddress)
+    const { start, end } = this.orderColumnRangeEnds(firstEnd, secondEnd)
 
-    if (start.sheet !== undefined && end.sheet === undefined) {
-      end = end.withAbsoluteSheet(start.sheet)
-    }
-
-    return buildColumnRangeAst(start, end, sheetReferenceType, range.leadingWhitespace)
+    return buildColumnRangeAst(start, end, sheetRefType, range.leadingWhitespace)
   })
+
   /**
    * Rule for row range, e.g. 1:2, Sheet1!1:2, Sheet1!1:Sheet1!2
    */
   private rowRangeExpression: AstRule = this.RULE('rowRangeExpression', () => {
-    const range = this.CONSUME(RowRange) as IExtendedToken
+    const range = this.CONSUME(RowRange) as ExtendedToken
     const [startImage, endImage] = range.image.split(':')
+    const firstAddress = this.ACTION(() => rowAddressFromString(this.sheetMapping, startImage, this.formulaAddress))
+    const secondAddress = this.ACTION(() => rowAddressFromString(this.sheetMapping, endImage, this.formulaAddress))
 
-    const start = this.ACTION(() => {
-      return rowAddressFromString(this.sheetMapping, startImage, this.formulaAddress)
-    })
-    let end = this.ACTION(() => {
-      return rowAddressFromString(this.sheetMapping, endImage, this.formulaAddress)
-    })
-
-    if (start === undefined || end === undefined) {
+    if (firstAddress === undefined || secondAddress === undefined) {
       return buildCellErrorAst(new CellError(ErrorType.REF))
     }
-    if (start.exceedsSheetSizeLimits(this.lexerConfig.maxRows) || end.exceedsSheetSizeLimits(this.lexerConfig.maxRows)) {
+
+    if (firstAddress.exceedsSheetSizeLimits(this.lexerConfig.maxRows) || secondAddress.exceedsSheetSizeLimits(this.lexerConfig.maxRows)) {
       return buildErrorWithRawInputAst(range.image, new CellError(ErrorType.NAME), range.leadingWhitespace)
     }
-    if (start.sheet === undefined && end.sheet !== undefined) {
+
+    if (firstAddress.sheet === undefined && secondAddress.sheet !== undefined) {
       return this.parsingError(ParsingErrorType.ParserError, 'Malformed range expression')
     }
 
-    const sheetReferenceType = this.rangeSheetReferenceType(start.sheet, end.sheet)
+    const { firstEnd, secondEnd, sheetRefType } = FormulaParser.fixSheetIdsForRangeEnds(firstAddress, secondAddress)
+    const { start, end } = this.orderRowRangeEnds(firstEnd, secondEnd)
 
-    if (start.sheet !== undefined && end.sheet === undefined) {
-      end = end.withAbsoluteSheet(start.sheet)
-    }
-
-    return buildRowRangeAst(start, end, sheetReferenceType, range.leadingWhitespace)
+    return buildRowRangeAst(start, end, sheetRefType, range.leadingWhitespace)
   })
+
   /**
    * Rule for cell reference expression (e.g. A1, $A1, A$1, $A$1, $Sheet42!A$17)
    */
   private cellReference: AstRule = this.RULE('cellReference', () => {
-    const cell = this.CONSUME(CellReference) as IExtendedToken
+    const cell = this.CONSUME(CellReference) as ExtendedToken
     const address = this.ACTION(() => {
       return cellAddressFromString(this.sheetMapping, cell.image, this.formulaAddress)
     })
@@ -272,8 +266,8 @@ export class FormulaParser extends EmbeddedActionsParser {
   /**
    * Rule for end range reference expression with additional checks considering range start
    */
-  private endRangeReference: AstRule = this.RULE('endRangeReference', (start: IExtendedToken) => {
-    const end = this.CONSUME(CellReference) as IExtendedToken
+  private endRangeReference: AstRule = this.RULE('endRangeReference', (start: ExtendedToken) => {
+    const end = this.CONSUME(CellReference) as ExtendedToken
 
     const startAddress = this.ACTION(() => {
       return cellAddressFromString(this.sheetMapping, start.image, this.formulaAddress)
@@ -300,7 +294,7 @@ export class FormulaParser extends EmbeddedActionsParser {
    *
    * End of range may be a cell reference or OFFSET() function call
    */
-  private endOfRangeExpression: AstRule = this.RULE('endOfRangeExpression', (start: IExtendedToken) => {
+  private endOfRangeExpression: AstRule = this.RULE('endOfRangeExpression', (start: ExtendedToken) => {
     return this.OR([
       {
         ALT: () => {
@@ -317,13 +311,7 @@ export class FormulaParser extends EmbeddedActionsParser {
             return buildCellErrorAst(new CellError(ErrorType.REF))
           }
           if (offsetProcedure.type === AstNodeType.CELL_REFERENCE) {
-            let end = offsetProcedure.reference
-            let sheetReferenceType = RangeSheetReferenceType.RELATIVE
-            if (startAddress.sheet !== undefined) {
-              sheetReferenceType = RangeSheetReferenceType.START_ABSOLUTE
-              end = end.withAbsoluteSheet(startAddress.sheet)
-            }
-            return buildCellRangeAst(startAddress, end, sheetReferenceType, start.leadingWhitespace?.image)
+            return this.buildCellRange(startAddress, offsetProcedure.reference, start.leadingWhitespace?.image)
           } else {
             return this.parsingError(ParsingErrorType.RangeOffsetNotAllowed, 'Range offset not allowed here')
           }
@@ -343,7 +331,7 @@ export class FormulaParser extends EmbeddedActionsParser {
    * Rule for end range reference expression starting with offset procedure with additional checks considering range start
    */
   private endRangeWithOffsetStartReference: AstRule = this.RULE('endRangeWithOffsetStartReference', (start: CellReferenceAst) => {
-    const end = this.CONSUME(CellReference) as IExtendedToken
+    const end = this.CONSUME(CellReference) as ExtendedToken
 
     const endAddress = this.ACTION(() => {
       return cellAddressFromString(this.sheetMapping, end.image, this.formulaAddress)
@@ -373,13 +361,7 @@ export class FormulaParser extends EmbeddedActionsParser {
         ALT: () => {
           const offsetProcedure = this.SUBRULE(this.offsetProcedureExpression)
           if (offsetProcedure.type === AstNodeType.CELL_REFERENCE) {
-            let end = offsetProcedure.reference
-            let sheetReferenceType = RangeSheetReferenceType.RELATIVE
-            if (start.reference.sheet !== undefined) {
-              sheetReferenceType = RangeSheetReferenceType.START_ABSOLUTE
-              end = end.withAbsoluteSheet(start.reference.sheet)
-            }
-            return buildCellRangeAst(start.reference, end, sheetReferenceType, start.leadingWhitespace)
+            return this.buildCellRange(start.reference, offsetProcedure.reference, start.leadingWhitespace)
           } else {
             return this.parsingError(ParsingErrorType.RangeOffsetNotAllowed, 'Range offset not allowed here')
           }
@@ -391,7 +373,7 @@ export class FormulaParser extends EmbeddedActionsParser {
    * Rule for expressions that start with OFFSET() function
    *
    * OFFSET() function can occur as cell reference or part of cell range.
-   * In order to preserve LL(k) properties, expressions that starts with OFFSET() functions needs to have separate rule.
+   * In order to preserve LL(k) properties, expressions that start with OFFSET() functions needs to have separate rule.
    *
    * Proper {@link Ast} node type is built depending on the presence of {@link RangeSeparator}
    */
@@ -440,18 +422,18 @@ export class FormulaParser extends EmbeddedActionsParser {
    * Rule for parenthesis expression
    */
   private parenthesisExpression: AstRule = this.RULE('parenthesisExpression', () => {
-    const lParenToken = this.CONSUME(LParen) as IExtendedToken
+    const lParenToken = this.CONSUME(LParen) as ExtendedToken
     const expression = this.SUBRULE(this.booleanExpression)
-    const rParenToken = this.CONSUME(RParen) as IExtendedToken
+    const rParenToken = this.CONSUME(RParen) as ExtendedToken
     return buildParenthesisAst(expression, lParenToken.leadingWhitespace, rParenToken.leadingWhitespace)
   })
   private arrayExpression: AstRule = this.RULE('arrayExpression', () => {
     return this.OR([
       {
         ALT: () => {
-          const ltoken = this.CONSUME(ArrayLParen) as IExtendedToken
+          const ltoken = this.CONSUME(ArrayLParen) as ExtendedToken
           const ret = this.SUBRULE(this.insideArrayExpression) as ArrayAst
-          const rtoken = this.CONSUME(ArrayRParen) as IExtendedToken
+          const rtoken = this.CONSUME(ArrayRParen) as ExtendedToken
           return buildArrayAst(ret.args, ltoken.leadingWhitespace, rtoken.leadingWhitespace)
         }
       },
@@ -475,7 +457,7 @@ export class FormulaParser extends EmbeddedActionsParser {
    * @param tokens - tokenized formula
    * @param formulaAddress - address of the cell in which formula is located
    */
-  public parseFromTokens(tokens: IExtendedToken[], formulaAddress: SimpleCellAddress): FormulaParserResult {
+  public parseFromTokens(tokens: ExtendedToken[], formulaAddress: SimpleCellAddress): FormulaParserResult {
     this.input = tokens
 
     let ast = this.formulaWithContext(formulaAddress)
@@ -544,19 +526,19 @@ export class FormulaParser extends EmbeddedActionsParser {
       },
       {
         ALT: () => {
-          const number = this.CONSUME(this.lexerConfig.NumberLiteral) as IExtendedToken
+          const number = this.CONSUME(this.lexerConfig.NumberLiteral) as ExtendedToken
           return buildNumberAst(this.numericStringToNumber(number.image), number.leadingWhitespace)
         },
       },
       {
         ALT: () => {
-          const str = this.CONSUME(StringLiteral) as IExtendedToken
+          const str = this.CONSUME(StringLiteral) as ExtendedToken
           return buildStringAst(str)
         },
       },
       {
         ALT: () => {
-          const token = this.CONSUME(ErrorLiteral) as IExtendedToken
+          const token = this.CONSUME(ErrorLiteral) as ExtendedToken
           const errString = token.image.toUpperCase()
           const errorType = this.lexerConfig.errorMapping[errString]
           if (errorType) {
@@ -573,7 +555,7 @@ export class FormulaParser extends EmbeddedActionsParser {
 
     const percentage = this.OPTION(() => {
       return this.CONSUME(PercentOp)
-    }) as Maybe<IExtendedToken>
+    }) as Maybe<ExtendedToken>
 
     if (percentage) {
       return buildPercentOpAst(positiveAtomicExpression, percentage.leadingWhitespace)
@@ -588,7 +570,7 @@ export class FormulaParser extends EmbeddedActionsParser {
     return this.OR([
       {
         ALT: () => {
-          const op = this.CONSUME(AdditionOp) as IExtendedToken
+          const op = this.CONSUME(AdditionOp) as ExtendedToken
           const value = this.SUBRULE(this.atomicExpression)
           if (tokenMatcher(op, PlusOp)) {
             return buildPlusUnaryOpAst(value, op.leadingWhitespace)
@@ -612,7 +594,7 @@ export class FormulaParser extends EmbeddedActionsParser {
     let lhs: Ast = this.SUBRULE(this.atomicExpression)
 
     this.MANY(() => {
-      const op = this.CONSUME(PowerOp) as IExtendedToken
+      const op = this.CONSUME(PowerOp) as ExtendedToken
       const rhs = this.SUBRULE2(this.atomicExpression)
 
       if (tokenMatcher(op, PowerOp)) {
@@ -633,7 +615,7 @@ export class FormulaParser extends EmbeddedActionsParser {
     let lhs: Ast = this.SUBRULE(this.powerExpression)
 
     this.MANY(() => {
-      const op = this.CONSUME(MultiplicationOp) as IExtendedToken
+      const op = this.CONSUME(MultiplicationOp) as ExtendedToken
       const rhs = this.SUBRULE2(this.powerExpression)
 
       if (tokenMatcher(op, TimesOp)) {
@@ -656,7 +638,7 @@ export class FormulaParser extends EmbeddedActionsParser {
     let lhs: Ast = this.SUBRULE(this.multiplicationExpression)
 
     this.MANY(() => {
-      const op = this.CONSUME(AdditionOp) as IExtendedToken
+      const op = this.CONSUME(AdditionOp) as ExtendedToken
       const rhs = this.SUBRULE2(this.multiplicationExpression)
 
       if (tokenMatcher(op, PlusOp)) {
@@ -679,7 +661,7 @@ export class FormulaParser extends EmbeddedActionsParser {
     let lhs: Ast = this.SUBRULE(this.additionExpression)
 
     this.MANY(() => {
-      const op = this.CONSUME(ConcatenateOp) as IExtendedToken
+      const op = this.CONSUME(ConcatenateOp) as ExtendedToken
       const rhs = this.SUBRULE2(this.additionExpression)
       lhs = buildConcatenateOpAst(lhs, rhs, op.leadingWhitespace)
     })
@@ -693,7 +675,7 @@ export class FormulaParser extends EmbeddedActionsParser {
     let lhs: Ast = this.SUBRULE(this.concatenateExpression)
 
     this.MANY(() => {
-      const op = this.CONSUME(BooleanOp) as IExtendedToken
+      const op = this.CONSUME(BooleanOp) as ExtendedToken
       const rhs = this.SUBRULE2(this.concatenateExpression)
 
       if (tokenMatcher(op, EqualsOp)) {
@@ -735,18 +717,64 @@ export class FormulaParser extends EmbeddedActionsParser {
     return this.formula()
   }
 
-  private buildCellRange(startAddress: CellAddress, endAddress: CellAddress, leadingWhitespace?: string): Ast {
-    if (startAddress.sheet === undefined && endAddress.sheet !== undefined) {
+  private buildCellRange(firstAddress: CellAddress, secondAddress: CellAddress, leadingWhitespace?: string): Ast {
+    if (firstAddress.sheet === undefined && secondAddress.sheet !== undefined) {
       return this.parsingError(ParsingErrorType.ParserError, 'Malformed range expression')
     }
 
-    const sheetReferenceType = this.rangeSheetReferenceType(startAddress.sheet, endAddress.sheet)
+    const { firstEnd, secondEnd, sheetRefType } = FormulaParser.fixSheetIdsForRangeEnds(firstAddress, secondAddress)
+    const { start, end } = this.orderCellRangeEnds(firstEnd, secondEnd)
 
-    if (startAddress.sheet !== undefined && endAddress.sheet === undefined) {
-      endAddress = endAddress.withAbsoluteSheet(startAddress.sheet)
+    return buildCellRangeAst(start, end, sheetRefType, leadingWhitespace)
+  }
+
+  private static fixSheetIdsForRangeEnds<T extends AddressWithSheet>(firstEnd: T, secondEnd: T): { firstEnd: T, secondEnd: T, sheetRefType: RangeSheetReferenceType } {
+    const sheetRefType = FormulaParser.rangeSheetReferenceType(firstEnd.sheet, secondEnd.sheet)
+    const secondEndFixed = (firstEnd.sheet !== undefined && secondEnd.sheet === undefined)
+      ? secondEnd.withSheet(firstEnd.sheet) as T
+      : secondEnd
+
+    return { firstEnd, secondEnd: secondEndFixed, sheetRefType }
+  }
+
+  private orderCellRangeEnds(endA: CellAddress, endB: CellAddress): { start: CellAddress, end: CellAddress } {
+    const ends = [ endA, endB ]
+    const [ startCol, endCol ] = ends.map(e => e.toColumnAddress()).sort(ColumnAddress.compareByAbsoluteAddress(this.formulaAddress))
+    const [ startRow, endRow ] = ends.map(e => e.toRowAddress()).sort(RowAddress.compareByAbsoluteAddress(this.formulaAddress))
+    const [ startSheet, endSheet ] = ends.map(e => e.sheet).sort(FormulaParser.compareSheetIds.bind(this))
+
+    return {
+      start: CellAddress.fromColAndRow(startCol, startRow, startSheet),
+      end: CellAddress.fromColAndRow(endCol, endRow, endSheet),
     }
+  }
 
-    return buildCellRangeAst(startAddress, endAddress, sheetReferenceType, leadingWhitespace)
+  private orderColumnRangeEnds(endA: ColumnAddress, endB: ColumnAddress): { start: ColumnAddress, end: ColumnAddress } {
+    const ends = [ endA, endB ]
+    const [ startCol, endCol ] = ends.sort(ColumnAddress.compareByAbsoluteAddress(this.formulaAddress))
+    const [ startSheet, endSheet ] = ends.map(e => e.sheet).sort(FormulaParser.compareSheetIds.bind(this))
+
+    return {
+      start: new ColumnAddress(startCol.type, startCol.col, startSheet),
+      end: new ColumnAddress(endCol.type, endCol.col, endSheet),
+    }
+  }
+
+  private orderRowRangeEnds(endA: RowAddress, endB: RowAddress): { start: RowAddress, end: RowAddress } {
+    const ends = [ endA, endB ]
+    const [ startRow, endRow ] = ends.sort(RowAddress.compareByAbsoluteAddress(this.formulaAddress))
+    const [ startSheet, endSheet ] = ends.map(e => e.sheet).sort(FormulaParser.compareSheetIds.bind(this))
+
+    return {
+      start: new RowAddress(startRow.type, startRow.row, startSheet),
+      end: new RowAddress(endRow.type, endRow.row, endSheet),
+    }
+  }
+  
+  private static  compareSheetIds(sheetA: number | undefined, sheetB: number | undefined): number {
+    sheetA = sheetA != null ? sheetA : Infinity
+    sheetB = sheetB != null ? sheetB : Infinity
+    return sheetA - sheetB
   }
 
   /**
@@ -849,7 +877,7 @@ export class FormulaParser extends EmbeddedActionsParser {
     return buildParsingErrorAst()
   }
 
-  private rangeSheetReferenceType(start?: number, end?: number): RangeSheetReferenceType {
+  private static rangeSheetReferenceType(start?: number, end?: number): RangeSheetReferenceType {
     if (start === undefined) {
       return RangeSheetReferenceType.RELATIVE
     } else if (end === undefined) {
@@ -860,10 +888,11 @@ export class FormulaParser extends EmbeddedActionsParser {
   }
 }
 
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
 type AstRule = (idxInCallingRule?: number, ...args: any[]) => (Ast)
 type OrArg = IOrAlt[] | OrMethodOpts
 
-export interface IExtendedToken extends IToken {
+export interface ExtendedToken extends IToken {
   leadingWhitespace?: IToken,
 }
 
@@ -891,7 +920,7 @@ export class FormulaLexer {
   }
 
   private skipWhitespacesInsideRanges(tokens: IToken[]): IToken[] {
-    return this.filterTokensByNeighbors(tokens, (previous: IToken, current: IToken, next: IToken) => {
+    return FormulaLexer.filterTokensByNeighbors(tokens, (previous: IToken, current: IToken, next: IToken) => {
       return (tokenMatcher(previous, CellReference) || tokenMatcher(previous, RangeSeparator))
         && tokenMatcher(current, this.lexerConfig.WhiteSpace)
         && (tokenMatcher(next, CellReference) || tokenMatcher(next, RangeSeparator))
@@ -899,14 +928,14 @@ export class FormulaLexer {
   }
 
   private skipWhitespacesBeforeArgSeparators(tokens: IToken[]): IToken[] {
-    return this.filterTokensByNeighbors(tokens, (previous: IToken, current: IToken, next: IToken) => {
+    return FormulaLexer.filterTokensByNeighbors(tokens, (previous: IToken, current: IToken, next: IToken) => {
       return !tokenMatcher(previous, this.lexerConfig.ArgSeparator)
         && tokenMatcher(current, this.lexerConfig.WhiteSpace)
         && tokenMatcher(next, this.lexerConfig.ArgSeparator)
     })
   }
 
-  private filterTokensByNeighbors(tokens: IToken[], shouldBeSkipped: (previous: IToken, current: IToken, next: IToken) => boolean): IToken[] {
+  private static filterTokensByNeighbors(tokens: IToken[], shouldBeSkipped: (previous: IToken, current: IToken, next: IToken) => boolean): IToken[] {
     if (tokens.length < 3) {
       return tokens
     }