+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- Add support to parse formulas with all whitespace characters (controlled by config param `ignoreWhiteSpace`). (#898)
+- Added support for reversed ranges (#834 & #938)
+- Added support for parsing formulas with all whitespace characters (controlled by config param `ignoreWhiteSpace`). (#898)
 
 ### Changed
 - **Breaking change**: Removed `gpu.js` dependency and its use. (#812)